cargo: bump to a minimal-versions compatible serde
The documentation for `serde`'s `derive` feature states that it should
really only be used for libraries with an optional `serde` feature
themselves. That's not happening here.

However, the need to split it is also required by `-Z minimal-versions`
builds. Without a direct dependency, `serde_derive` 1.0.0 gets chosen.
This is not suitable for this crate because it uses the `alias` feature.
1.0.103 is selected because it is the version that is compatible with
`-Z minimal-versions`.