normalize dir seperators in types
Rust anonymous closure types include the path to the position in the file
where the closure was defined. This results in different output on windows
compared to linux and mac. This patch looks for lines starting with
"= note: required because it appears within the type", and replaces all
'\' with '/' in order to be cross platform.