Use the `Cargo.lock` from the workspace
This changes the way trybuild is generating the `Cargo.lock`. Instead of
generating a fresh `Cargo.lock`, it takes the `Cargo.lock` of the
workspace as "input". This should prevent compile problems because crate
patch versions are bumped on crates.io, but cargo fails to resolve a
proper `Cargo.lock` file. By using the workspace `Cargo.lock` we can be
sure that a working `Cargo.lock` file is used.