Use subdirectory within target/tests/
Or to put it another way, do not assume that the project's
target/tests/target/ is ours to use as we wish.

Empirically, this is sufficient on its own to resolve
  https://github.com/eupn/macrotest/issues/83
  https://github.com/dtolnay/trybuild/issues/218

(I don't know precisely what the difference is between trybuild's and
mcarotest's build invocations.  It is possible that it would be better
for cargo to be able to cache both in the target directory, but I
think we should use a namespaced directory, regardless.)