Update MSRV to reflect changes to dependencies
Replacing basic-toml crate in dtolnay/trybuild#239 raised MSRV
to 1.67 (toml_datetime specifies this minimum), but CI workflow was
updated to 1.70 in that PR, so this patch raises MSRV to 1.70
instead.

A reference to even older MSRV in the README is removed entirely
since this information lives in Cargo.toml as of Rust 1.56.